Move utility functions to separate file
