+++ utils.py
@@ -1,0 +1,37 @@
+import io
+import os
+import zipfile
+
+import cv2
+import numpy as np
+from fastapi import Response
+
+
+def load_image_into_numpy_array(data):
+    npimg = np.frombuffer(data, np.uint8)
+    frame = cv2.imdecode(npimg, cv2.IMREAD_COLOR)
+    return frame
+
+
+def zipfiles(filenames: list):
+    zip_filename = "archive.zip"
+
+    s = io.BytesIO()
+    zf = zipfile.ZipFile(s, "w")
+
+    for fpath in filenames:
+        # Calculate path for file in zip
+        fdir, fname = os.path.split(fpath)
+
+        # Add file, at correct path
+        zf.write(fpath, fname)
+
+    # Must close zip for all contents to be written
+    zf.close()
+
+    # Grab ZIP file from in-memory, make response with correct MIME-type
+    resp = Response(s.getvalue(), media_type="application/x-zip-compressed", headers={
+        'Content-Disposition': f'attachment;filename={zip_filename}'
+    })
+
+    return resp
